+++ projects/pastanaga-angular/src/lib/modal/modal.model.ts
@@ -2,6 +2,7 @@ import { Observable, Subject } from 'rxjs';
 
 export interface IModalConfig<D = any> {
     blocking?: boolean;
+    fullscreen?: boolean;
     showOverlay?: boolean;
     withCloseButton?: boolean;
     data?: D;
@@ -9,6 +10,7 @@ export interface IModalConfig<D = any> {
 
 export class ModalConfig<D = any> {
     blocking: boolean;
+    fullscreen: boolean;
     showOverlay: boolean;
     withCloseButton: boolean;
     readonly data?: Readonly<D>;
@@ -17,6 +19,7 @@ export class ModalConfig<D = any> {
         this.blocking = !!config && typeof config.blocking === 'boolean' ? config.blocking : true;
         this.showOverlay = !!config && typeof config.showOverlay === 'boolean' ? config.showOverlay : true;
         this.withCloseButton = !!config && typeof config.withCloseButton === 'boolean' ? config.withCloseButton : false;
+        this.fullscreen = !!config && typeof config.fullscreen === 'boolean' ? config.fullscreen : false;
         this.data = config?.data;
     }
 }